Move AppleWin to the Windows folder.
