+++ source/Configuration/PageConfig.cpp
@@ -26,7 +26,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "PageConfig.h"
 #include "PropertySheetHelper.h"
 
-#include "../AppleWin.h"
+#include "../Windows/AppleWin.h"
 #include "../Windows/WinFrame.h"
 #include "../Registry.h"
 #include "../SerialComms.h"
