+++ source/Configuration/PageDisk.cpp
@@ -26,7 +26,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "PageDisk.h"
 #include "PropertySheetHelper.h"
 
-#include "../AppleWin.h"
+#include "../Windows/AppleWin.h"
 #include "../CardManager.h"
 #include "../Disk.h"	// Drive_e, Disk_Status_e
 #include "../Windows/WinFrame.h"
