+++ source/Configuration/PropertySheet.cpp
@@ -31,7 +31,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 #include "PropertySheet.h"
 
-#include "../AppleWin.h"
+#include "../Windows/AppleWin.h"
 #include "../Windows/WinFrame.h"
 #include "../resource/resource.h"
