+++ source/Debugger/Debugger_Display.cpp
@@ -32,7 +32,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "Debug.h"
 #include "Debugger_Display.h"
 
-#include "../AppleWin.h"
+#include "../Windows/AppleWin.h"
 #include "../Core.h"
 #include "../CPU.h"
 #include "../Frame.h"
