+++ source/Debugger/Debugger_Symbols.cpp
@@ -30,7 +30,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 #include "Debug.h"
 
-#include "../AppleWin.h"
+#include "../Windows/AppleWin.h"
 #include "../Core.h"
 
 	// 2.6.2.13 Added: Can now enable/disable selected symbol table(s) !
