+++ source/Joystick.cpp
@@ -40,7 +40,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "Joystick.h"
-#include "AppleWin.h"
+#include "Windows/AppleWin.h"
 #include "CPU.h"
 #include "Memory.h"
 #include "YamlHelper.h"
