+++ source/Keyboard.cpp
@@ -29,7 +29,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "Keyboard.h"
-#include "AppleWin.h"
+#include "Windows/AppleWin.h"
 #include "Core.h"
 #include "Windows/WinFrame.h"
 #include "Pravets.h"
