+++ source/NTSC_CharSet.cpp
@@ -21,7 +21,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 #include "StdAfx.h"
 #include "NTSC_CharSet.h"
-#include "AppleWin.h"
+#include "Windows/AppleWin.h"
 #include "Core.h"
 #include "../resource/resource.h"
 #include "Video.h"
