+++ source/SaveState.cpp
@@ -31,7 +31,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "SaveState.h"
 #include "YamlHelper.h"
 
-#include "AppleWin.h"
+#include "Windows/AppleWin.h"
 #include "CardManager.h"
 #include "CPU.h"
 #include "Debug.h"
