+++ source/Windows/AppleWin.cpp
@@ -28,7 +28,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 #include "StdAfx.h"
 
-#include "AppleWin.h"
+#include "Windows/AppleWin.h"
 #include "Utilities.h"
 #include "Debug.h"
 #include "Keyboard.h"
