+++ source/Windows/WinFrame.cpp
@@ -29,7 +29,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "Windows/WinFrame.h"
-#include "AppleWin.h"
+#include "Windows/AppleWin.h"
 #include "CardManager.h"
 #include "CPU.h"
 #include "Disk.h"
