+++ source/Windows/WinVideo.cpp
@@ -30,7 +30,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 #include "Windows/WinVideo.h"
 #include "Windows/WinFrame.h"
-#include "AppleWin.h"
+#include "Windows/AppleWin.h"
 #include "Video.h"
 #include "Core.h"
 #include "CPU.h"
