+++ test/TestCPU6502/TestCPU6502.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
 
-#include "../../source/AppleWin.h"
+#include "../../source/Windows/AppleWin.h"
 #include "../../source/CPU.h"
 #include "../../source/Memory.h"
 #include "../../source/SynchronousEventManager.h"
